backend/requirements.txt
Django>=3.2,<4.0
clorm==1.4.2
djangorestframework==3.14.0
django-cors-headers == 4.3.1
